+++ lib/src/widgets/chat/typing/ds_typing_message_bubble.widget.dart
@@ -1,6 +1,8 @@
 import 'package:blip_ds/blip_ds.dart';
 import 'package:flutter/material.dart';
 
+/// Creates an animation showing that someone is typing something by responding to an interaction
+///
 /// Use [align] left or right to position typing on screen
 class DSTypingMessageBubble extends StatelessWidget {
   final DSAlign align;