+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.0.12
 
-* Add typing message bubble
+* Add button styles
+* Add unsupported content bubble
+* Add file message bubble
+* Add typing dots message bubble
 * Add animated dots
 
 ## 0.0.11